move aircraft_performance.py from models to core and update imports accordingly
